housekeeping; fix #466
