+++ game_view.cpp
@@ -5,6 +5,7 @@
 #include "food.h"
 #include "game.h"
 #include "game_resources.h"
+#include "game_functions.h"
 #include <SFML/Graphics.hpp>
 #include <SFML/Graphics/Text.hpp>
 #include <cmath>
